requirements.txt
h5py==3.2.1
matplotlib==3.4.2
numpy==1.19.5
Pillow==9.1.0
PyYAML==6.0
scipy==1.6.3
setuptools==50.3.1
SharedArray==3.2.1
tensorboardX==2.5
termcolor==1.1.0
timm==0.4.9
torch==1.7.1
torch_geometric==1.7.0
torch_points3d==1.3.0
torch_points_kernels==0.6.10
torch_scatter==2.0.6
torchvision==0.8.2
